+++ src/apsg/plotting/_styles.py
@@ -110,17 +110,17 @@ class StereoNetConeStyle(StereoNetStyle):
         return StereoNetArtistFactory.create_cone(*filtered, **self.kwargs)
 
 
-class StereoNetBinghamStyle(StereoNetStyle):
+class StereoNetConfidenceStyle(StereoNetStyle):
     def __init__(self, **kwargs):
-        """Initialize stereonet Bingham confidence ellipse style."""
+        """Initialize stereonet confidence cone/ellipse style."""
         self._valid = Vector3Set
-        self.kwargs = getattr(apsg_conf, "stereonet_bingham").copy()
+        self.kwargs = getattr(apsg_conf, "stereonet_confidence").copy()
         super().__init__(**kwargs)
 
     def create_artist(self, *args):
-        """Create stereonet Bingham confidence ellipse artist with configured style."""
+        """Create stereonet confidence cone/ellipse artist with configured style."""
         filtered = (arg for arg in args if isinstance(arg, self._valid))
-        return StereoNetArtistFactory.create_bingham(*filtered, **self.kwargs)
+        return StereoNetArtistFactory.create_confidence(*filtered, **self.kwargs)
 
 
 class StereoNetPairStyle(StereoNetStyle):
@@ -248,9 +248,9 @@ class StereoNetStyleFactory:
         return StereoNetConeStyle(**kwargs)
 
     @staticmethod
-    def bingham(**kwargs):
-        """Return StereoNetBinghamStyle with given kwargs."""
-        return StereoNetBinghamStyle(**kwargs)
+    def confidence(**kwargs):
+        """Return StereoNetConfidenceStyle with given kwargs."""
+        return StereoNetConfidenceStyle(**kwargs)
 
     @staticmethod
     def pair(**kwargs):